Extract opengl pipeline into its own class
