+++ GesturesToInputs/main.cpp
@@ -1,7 +1,3 @@
-#include "glad/glad.h"
-#define GLFW_INCLUDE_NONE
-#include "glfw/include/GLFW/glfw3.h"
-
 #include "main.h"
 
 #include <dinput.h>
@@ -10,8 +6,10 @@
 #include <thread>
 #include <vector>
 
+#include "structures.h"
 #include "Gesture.h"
 #include "GesturesToInputsProcessor_CPU.h"
+#include "GesturesToInputsPorcessor_OpenGL.h"
 #include "MediaFoundationWebcam.h"
 #include "Shader.h"
 
@@ -146,356 +144,6 @@ std::vector<GestureInput> testGestures() {
     };
 }
 
-GLFWwindow* window;
-
-struct ThresholdData {
-    float lowColour[4];
-    float highColour[4];
-    float trackerColour[4];
-
-    ThresholdData(float _lowColour[4], float _highColour[4], float _trackerColour[4]) {
-        lowColour[0] = _lowColour[0];
-        lowColour[1] = _lowColour[1];
-        lowColour[2] = _lowColour[2];
-        lowColour[3] = _lowColour[3];
-        highColour[0] = _highColour[0];
-        highColour[1] = _highColour[1];
-        highColour[2] = _highColour[2];
-        highColour[3] = _highColour[3];
-        trackerColour[0] = _trackerColour[0];
-        trackerColour[1] = _trackerColour[1];
-        trackerColour[2] = _trackerColour[2];
-        trackerColour[3] = _trackerColour[3];
-    }
-};
-
-struct ObjectSearchData {
-    GLuint boundingBox[4];
-    GLuint objectTrackerIndex;
-    GLuint padding[3];
-};
-
-struct TrackerData {
-    float colour[4];
-    uint area = 0;
-    uint horizontalPosition = 0;
-    uint verticalPosition = 0;
-    uint orientation = 0;
-    uint isVisibleInFrame = 0;
-    uint padding[3];
-    
-    TrackerData(float _colour[4]) {
-        colour[0] = _colour[0];
-        colour[1] = _colour[1];
-        colour[2] = _colour[2];
-        colour[3] = _colour[3];
-        padding[0] = padding[1] = padding[2] = 0;
-    }
-};
-
-struct GestureRuleData {
-    uint gestureIndex = 0;
-    uint type = 0;
-    int operation = 0;
-    uint expectedValue = 0;
-    uint compareTwoTrackers = 0;
-    uint trackerIndex = 0;
-    uint comparisonTrackerIndex = 0;
-    uint padding = 0;
-};
-
-typedef cv::Vec<uint, 4> Vec4ui;
-
-static const struct
-{
-    float x, y;
-    float xTexture, yTexture;
-} vertices[4] =
-{
-    { -1.0f,  1.0f, 0.0f, 1.0f },
-    { -1.0f, -1.0f, 0.0f, 0.0f },
-    {  1.0f, -1.0f, 1.0f, 0.0f },
-    {  1.0f,  1.0f, 1.0f, 1.0f },
-};
-
-const unsigned int indices[] = { 0, 1, 2, 3, 0, 2 };
-const int INPUT_IMAGE_UNIT = 0;
-const int THRESHOLD_IMAGE_UNIT = 1;
-const int OUTPUT_IMAGE_UNIT = 2;
-const int SHADER_STORAGE_THRESHOLD = 3;
-const int SHADER_STORAGE_OBJECT_SEARCH = 4;
-const int RAW_DATA_IMAGE_UNIT = 5;
-const int SHADER_STORAGE_TRACKERS = 6;
-const int SHADER_STORAGE_GESTURE = 7;
-const int SHADER_STORAGE_GESTURE_RULE = 8;
-
-const unsigned int sampleColumns = 64;
-const unsigned int sampleRows = 48;
-const unsigned int totalSamples = sampleColumns * sampleRows;
-
-bool opengl_has_errored = false;
-GLuint vertex_array;
-GLuint vertex_buffer, element_buffer;
-GLint vpos_location, tex_location;
-
-GLuint inputTextureHandle, thresholdTextureHandle, outputTextureHandle, rawDataTextureHandle;
-const GLenum inputTextureUnit = GL_TEXTURE0, thresholdTextureUnit = GL_TEXTURE1, outputTextureUnit = GL_TEXTURE2, rawDataTextureUnit = GL_TEXTURE3;
-int xMaxInstances, yMaxInstances, zMaxInstances, totalMaxInstances;
-
-GLuint thresholdShaderBuffer, trackerShaderBuffer, computeShaderBuffer, gestureFoundShaderBuffer, gestureRuleShaderBuffer;
-
-Shader hsvShader, thresholdShader, objectSearchShader, detectGesturesShader, displayShader, yuy2Shader;
-
-cv::Mat source;
-unsigned int sourceWidth, sourceHeight;
-
-void error_callback(int error, const char* description)
-{
-    fprintf(stderr, "GLFW Error: %s\n", description);
-}
-
-int end(std::string message = "") {
-    GLuint textures[] = { inputTextureHandle, thresholdTextureHandle, outputTextureHandle, rawDataTextureHandle };
-    glDeleteTextures(sizeof(textures) / sizeof(textures[0]), &textures[0]);
-
-    glBindBuffer(GL_SHADER_STORAGE_BUFFER, 0);
-    glBindBuffer(GL_ARRAY_BUFFER, 0);
-    glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
-    
-    GLuint buffers[] = { thresholdShaderBuffer, trackerShaderBuffer, computeShaderBuffer, gestureFoundShaderBuffer, gestureRuleShaderBuffer, vertex_buffer, element_buffer };
-    glDeleteBuffers(sizeof(buffers) / sizeof(buffers[0]), &buffers[0]);
-    
-    glDeleteVertexArrays(1, &vertex_array);
-
-    glDeleteProgram(hsvShader.getHandle());
-    glDeleteProgram(thresholdShader.getHandle());
-    glDeleteProgram(objectSearchShader.getHandle());
-    glDeleteProgram(detectGesturesShader.getHandle());
-    glDeleteProgram(displayShader.getHandle());
-    glDeleteProgram(yuy2Shader.getHandle());
-
-    glfwDestroyWindow(window);
-    glfwTerminate();
-    std::cout << message << std::endl;
-    return 1;
-}
-
-static void key_callback(GLFWwindow* window, int key, int scancode, int action, int mods)
-{
-    if (key == GLFW_KEY_ESCAPE && action == GLFW_PRESS)
-        glfwSetWindowShouldClose(window, GLFW_TRUE);
-}
-
-static void MessageCallback(GLenum source, GLenum type, GLuint id, GLenum severity, GLsizei length, const GLchar* message, const void* userParam) {
-    opengl_has_errored = true;
-    fprintf(stderr, "GL CALLBACK: %s type = 0x%x, severity = 0x%x, message = %s\n",
-        (type == GL_DEBUG_TYPE_ERROR ? "** GL ERROR **" : ""),
-        type, severity, message);
-}
-
-int setup() {
-    // TODO:: If no GPU or opengl version < 4.3 then fall back to cpu method.
-    // TODO:: Cpu method may be optimisable since we only need rough bounding box. Moments can certainly be removed, and perhaps more optimisaitons.
-
-    if (!glfwInit())
-    {
-        // Initialization failed
-        return end("GLFW could not initialise OpenGL");
-    }
-
-    glfwSetErrorCallback(error_callback);
-    glfwWindowHint(GLFW_OPENGL_DEBUG_CONTEXT, GL_TRUE);
-    glfwWindowHint(GLFW_CONTEXT_ROBUSTNESS, GLFW_LOSE_CONTEXT_ON_RESET);
-
-    glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 4);
-    glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 4);
-    glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);
-
-    window = glfwCreateWindow(sourceWidth, sourceHeight, "Tracked Objects", NULL, NULL);
-    if (!window)
-    {
-        return end("GLFW could not create a window");
-    }
-    
-    glfwSetKeyCallback(window, key_callback);
-    glfwMakeContextCurrent(window);
-    
-    if (!gladLoadGLLoader((GLADloadproc)glfwGetProcAddress))
-    {
-        return end("Could not set up GLAD");
-    }
-
-    glDebugMessageCallback(MessageCallback, NULL);
-    glDebugMessageControl(GL_DONT_CARE, GL_DONT_CARE, GL_DONT_CARE, 0, NULL, GL_FALSE);
-    glDebugMessageControl(GL_DONT_CARE, GL_DONT_CARE, GL_DEBUG_SEVERITY_HIGH, 0, NULL, GL_TRUE);
-    glDebugMessageControl(GL_DONT_CARE, GL_DONT_CARE, GL_DEBUG_SEVERITY_MEDIUM, 0, NULL, GL_TRUE);
-
-    glfwSwapInterval(0);
-
-    glClearColor(0.2f, 0.3f, 0.3f, 1.0f);
-
-    int work_group_size[3];
-    glGetIntegeri_v(GL_MAX_COMPUTE_WORK_GROUP_SIZE, 0, &work_group_size[0]);
-    glGetIntegeri_v(GL_MAX_COMPUTE_WORK_GROUP_SIZE, 1, &work_group_size[1]);
-    glGetIntegeri_v(GL_MAX_COMPUTE_WORK_GROUP_SIZE, 2, &work_group_size[2]);
-    xMaxInstances = work_group_size[0];
-    yMaxInstances = work_group_size[1];
-    zMaxInstances = work_group_size[2];
-    glGetIntegerv(GL_MAX_COMPUTE_WORK_GROUP_INVOCATIONS, &totalMaxInstances);
-
-    std::cout << std::endl << "(" << xMaxInstances << ", " << yMaxInstances << ", " << zMaxInstances << ") = " << totalMaxInstances << std::endl;
-
-    int maxUnits;
-    glGetIntegerv(GL_MAX_IMAGE_UNITS, &maxUnits);
-    int maxComputeUnits;
-    glGetIntegerv(GL_MAX_COMPUTE_IMAGE_UNIFORMS, &maxComputeUnits);
-    std::cout << "Units " << maxUnits << " & Uniforms " << maxComputeUnits << std::endl;
-
-    return 0;
-}
-
-void bindImageHandle(GLuint* handle, GLenum textureUnit, int binding, int format = GL_RGBA32F, int access = GL_READ_WRITE) {
-    glActiveTexture(textureUnit);
-    glGenTextures(1, handle);
-    glBindTexture(GL_TEXTURE_2D, *handle);
-    glTexStorage2D(GL_TEXTURE_2D, 1, format, sourceWidth, sourceHeight);
-
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_REPEAT);
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_REPEAT);
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR_MIPMAP_LINEAR);
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-    glBindImageTexture(binding, *handle, 0, GL_FALSE, 0, access, format);
-}
-
-void convertYUY2ToRGB() {
-    yuy2Shader.compileCompute("Convert_YUY2.comp", sourceWidth / 2, sourceHeight, 1);
-    yuy2Shader.setUniform("inputImage", INPUT_IMAGE_UNIT);
-    yuy2Shader.setUniform("rawData", RAW_DATA_IMAGE_UNIT);
-    bindImageHandle(&rawDataTextureHandle, rawDataTextureUnit, RAW_DATA_IMAGE_UNIT, GL_RG8UI, GL_READ_ONLY);
-}
-
-void convertToHSV() {
-    hsvShader.compileCompute("ConvertToHSV.comp", sourceWidth, sourceHeight, 1);
-
-    hsvShader.setUniform("inputImage", INPUT_IMAGE_UNIT);
-    hsvShader.setUniform("thresholdTexture", THRESHOLD_IMAGE_UNIT);
-    hsvShader.setUniform("outputImage", OUTPUT_IMAGE_UNIT);
-    hsvShader.setUniform("inputIsMirrored", unsigned int(1));
-    hsvShader.setUniform3("rgbCorrections", new float[3]{ 0.0f, 0.05f, 0.0f });
-    hsvShader.setUniform("lightenAmount", 0.5f);
-
-    bindImageHandle(&thresholdTextureHandle, thresholdTextureUnit, THRESHOLD_IMAGE_UNIT);
-    bindImageHandle(&outputTextureHandle, outputTextureUnit, OUTPUT_IMAGE_UNIT);
-}
-
-template<typename T>void createStorageBuffer(GLuint* bufferHandle, std::vector<T>* items, int binding) {
-    glGenBuffers(1, bufferHandle);
-    glBindBuffer(GL_SHADER_STORAGE_BUFFER, *bufferHandle);
-    glBindBufferBase(GL_SHADER_STORAGE_BUFFER, binding, *bufferHandle);
-
-    uint dataSize = sizeof(T) * items->size();
-    T* bufferItems;
-    bufferItems = (T*)malloc(dataSize);
-    int i = 0;
-    for (auto& item : *items) { bufferItems[i] = item; i++; }
-    glBufferData(GL_SHADER_STORAGE_BUFFER, dataSize, bufferItems, GL_STATIC_DRAW);
-    free(bufferItems);
-
-}
-
-void createEmptyStorageBuffer(GLuint* bufferHandle, unsigned int dataSize, int binding) {
-    glGenBuffers(1, bufferHandle);
-    glBindBuffer(GL_SHADER_STORAGE_BUFFER, *bufferHandle);
-    glBindBufferBase(GL_SHADER_STORAGE_BUFFER, binding, *bufferHandle);
-    glBufferData(GL_SHADER_STORAGE_BUFFER, dataSize, NULL, GL_STATIC_COPY);
-}
-
-unsigned int* createEmptyPersistantStorageBuffer(GLuint* bufferHandle, unsigned int dataSize, int binding) {
-    auto flags = GL_MAP_READ_BIT | GL_MAP_PERSISTENT_BIT;
-    glGenBuffers(1, bufferHandle);
-    glBindBuffer(GL_SHADER_STORAGE_BUFFER, *bufferHandle);
-    glBindBufferBase(GL_SHADER_STORAGE_BUFFER, binding, *bufferHandle);
-    glBufferStorage(GL_SHADER_STORAGE_BUFFER, dataSize, NULL, flags);
-    return (unsigned int*)glMapBufferRange(GL_SHADER_STORAGE_BUFFER, 0, dataSize, flags);
-}
-
-void threshold(std::vector<ThresholdData>* items) {
-    thresholdShader.compileCompute("Threshold.comp", sourceWidth, sourceHeight, 1);
-    thresholdShader.setUniform("numberOfColours", unsigned int(items->size()));
-    thresholdShader.setUniform("thresholdTexture", THRESHOLD_IMAGE_UNIT);
-    createStorageBuffer(&thresholdShaderBuffer, items, SHADER_STORAGE_THRESHOLD);
-}
-
-void searchForObjects(std::vector<TrackerData>* trackers) {
-    objectSearchShader.compileCompute("ObjectBoundingBoxSearch_Pass1.comp", sampleColumns, sampleRows, trackers->size());
-
-    objectSearchShader.setUniform("thresholdTexture", THRESHOLD_IMAGE_UNIT);
-
-    objectSearchShader.setUniform2("samplePixelDimensions", new unsigned int[2] { sourceWidth / sampleColumns, sourceHeight / sampleRows });
-    objectSearchShader.setUniform2("sampleSize", new unsigned int [2] { sampleColumns, sampleRows });
-    objectSearchShader.setUniform("numberOfPasses", unsigned int(100));
-    objectSearchShader.setUniform("threshold", unsigned int(500));
-    objectSearchShader.setUniform("outputImage", OUTPUT_IMAGE_UNIT);
-    
-    createEmptyStorageBuffer(&computeShaderBuffer, sizeof(ObjectSearchData) * totalSamples, SHADER_STORAGE_OBJECT_SEARCH);
-    createStorageBuffer(&trackerShaderBuffer, trackers, SHADER_STORAGE_TRACKERS);
-}
-
-unsigned int* detectGesturesSetup(unsigned int numberOfGestures, std::vector<GestureRuleData>* rules) {
-    detectGesturesShader.compileCompute("DetectGestures.comp", rules->size(), 1, 1);
-    auto gestureFoundDataPointer = createEmptyPersistantStorageBuffer(&gestureFoundShaderBuffer, numberOfGestures * sizeof(unsigned int), SHADER_STORAGE_GESTURE);
-    createStorageBuffer(&gestureRuleShaderBuffer, rules, SHADER_STORAGE_GESTURE_RULE);
-    return gestureFoundDataPointer;
-}
-
-void displayOutputSetup() {
-    glGenVertexArrays(1, &vertex_array);
-    glBindVertexArray(vertex_array);
-
-    glGenBuffers(1, &vertex_buffer);
-    glBindBuffer(GL_ARRAY_BUFFER, vertex_buffer);
-    glBufferData(GL_ARRAY_BUFFER, sizeof(vertices), vertices, GL_STATIC_DRAW);
-
-    glGenBuffers(1, &element_buffer);
-    glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, element_buffer);
-    glBufferData(GL_ELEMENT_ARRAY_BUFFER, sizeof(indices), indices, GL_STATIC_DRAW);
-
-    displayShader.addShader(GL_VERTEX_SHADER, "test.vert");
-    displayShader.addShader(GL_FRAGMENT_SHADER, "test.frag");
-    displayShader.compile();
-    displayShader.use();
-
-    vpos_location = displayShader.attributeLocation("aPosition");
-    glVertexAttribPointer(vpos_location, 2, GL_FLOAT, GL_FALSE, sizeof(vertices[0]), (void*)0);
-
-    tex_location = displayShader.attributeLocation("aTexCoord");
-    glVertexAttribPointer(tex_location, 2, GL_FLOAT, GL_FALSE, sizeof(vertices[0]), (void*)(2 * sizeof(float)));
-
-    displayShader.setUniform("outputImage", OUTPUT_IMAGE_UNIT);
-
-    glViewport(0, 0, sourceWidth, sourceHeight);
-}
-
-void displayOutput() {
-    displayShader.use();
-    glEnableVertexAttribArray(vpos_location);
-    glEnableVertexAttribArray(tex_location);
-    glDrawElements(GL_TRIANGLES, 6, GL_UNSIGNED_INT, (void*)0);
-    glDisableVertexAttribArray(vpos_location);
-    glDisableVertexAttribArray(tex_location);
-    glfwSwapBuffers(window);
-}
-
-void debugDisplayTexture(GLenum textureUnit, std::string windowName) {
-    glActiveTexture(textureUnit);
-    unsigned char* gl_texture_bytes = (unsigned char*)malloc(sizeof(unsigned char) * sourceWidth * sourceHeight * 3);
-    glGetTexImage(GL_TEXTURE_2D, 0, GL_BGR, GL_UNSIGNED_BYTE, gl_texture_bytes);
-    cv::imshow(windowName, cv::Mat(sourceHeight, sourceWidth, CV_8UC3, gl_texture_bytes));
-    free(gl_texture_bytes);
-}
-
 std::vector<GestureRuleData>* getRulesFromGestures(std::vector<GestureInput>* gestures, std::map<std::string, unsigned int>* trackerNameIndexMap) {
     auto rules = new std::vector<GestureRuleData>;
     unsigned int gestureIndex = 0;
@@ -526,77 +174,35 @@ std::vector<ThresholdData>* getThresholdData(float greenTracker[4], float redTra
 
 int main(int argc, char** argv)
 {
+    // TODO:: If no GPU or opengl version < 4.3 then fall back to cpu method.
+    // TODO:: Cpu method may be optimisable since we only need rough bounding box. Moments can certainly be removed, and perhaps more optimisaitons.
+
     bool useGPU = true;
-    Shader::setRoot(std::filesystem::path(argv[0]).parent_path() / "shaders");
 
     if (useGPU) {
-        MediaFoundationWebcam* webcam = new MediaFoundationWebcam();
-        std::thread t1(&MediaFoundationWebcam::CreateVideoCaptureDevice, webcam);
-        t1.detach();
+        try {
+            Shader::setRoot(std::filesystem::path(argv[0]).parent_path() / "shaders");
+            auto gestures = testGestures();
+            float greenTracker[4] = { 87, 183, 183, 255 }; float redTracker[4] = { 174, 179, 205, 255 };
+            
+            MediaFoundationWebcam* webcam = new MediaFoundationWebcam();
+            std::thread t1(&MediaFoundationWebcam::CreateVideoCaptureDevice, webcam);
+            t1.detach();
 
-        webcam->wait();
-        sourceWidth = webcam->getWidth(); sourceHeight = webcam->getHeight();
-
-        int status = setup();
-        if (status != 0) { return end(); }
-
-        float greenTracker[4] = { 87, 183, 183, 255 }; float redTracker[4] = { 174, 179, 205, 255 };
-        auto gesture = Gesture::getInstance();
-        auto gestures = testGestures();
-
-        PerformanceTimer perf;
-        bindImageHandle(&inputTextureHandle, inputTextureUnit, INPUT_IMAGE_UNIT);
-        convertYUY2ToRGB();
-        convertToHSV();
-        threshold(getThresholdData(greenTracker, redTracker));
-        searchForObjects(new std::vector<TrackerData>{ TrackerData(greenTracker), TrackerData(redTracker) });
-        auto gestureFoundDataPointer = detectGesturesSetup(gestures.size(), getRulesFromGestures(&gestures, new std::map<std::string, unsigned int>{ {"Green", 0}, {"Red", 1} }));
-        displayOutputSetup();
-
-        // TODO:: Allow more formats. If YUY2 it can be converted, if it's a format GLTexSubImage2D knows about it can be uploaded directly
-        while (!glfwWindowShouldClose(window) && !opengl_has_errored)
-        {
             webcam->wait();
+            auto openGlProcessor = new GesturesToInputsPorcessor_OpenGL(
+                webcam,
+                getThresholdData(greenTracker, redTracker),
+                new std::vector<TrackerData>{ TrackerData(greenTracker), TrackerData(redTracker) },
+                getRulesFromGestures(&gestures, new std::map<std::string, unsigned int>{ {"Green", 0}, {"Red", 1} }),
+                &gestures
+            );
 
-            perf.Start();
-            auto bytes = webcam->getData();
-            glActiveTexture(rawDataTextureUnit);
-            glTexSubImage2D(GL_TEXTURE_2D, 0, 0, 0, sourceWidth, sourceHeight, GL_RG_INTEGER, GL_UNSIGNED_BYTE, bytes);
-            glMemoryBarrier(GL_ALL_BARRIER_BITS);
-
-            yuy2Shader.compute();
-            hsvShader.compute();
-            thresholdShader.compute();
-            //debugDisplayTexture(thresholdTextureUnit, "threshold");
-            objectSearchShader.compute();
-            detectGesturesShader.compute();
-
-            std::cout << "Processed Frame: ";  perf.End();
-
-            perf.Start();
-            gesture->reset();
-            for (unsigned int i = 0; i < gestures.size(); i++) {
-                gesture->handleInput(&gestures.at(i), gestureFoundDataPointer[i] == 1);
-            }
-            gesture->complete();
-
-            std::cout << "Processed Gestures: ";  perf.End(); std::cout << std::endl;
-
-            perf.Start();
-            displayOutput();
-            std::cout << "Output displayed in "; perf.End();
-            std::cout << std::endl << std::endl << std::endl;
-
-            glfwPollEvents();
+            openGlProcessor->run();
         }
-
-        glBindBuffer(GL_SHADER_STORAGE_BUFFER, gestureFoundShaderBuffer);
-        glUnmapBuffer(GL_SHADER_STORAGE_BUFFER);
-
-        webcam->Close();
-        webcam->Release();
-
-        return end();
+        catch (std::string e) {
+            std::cout << "An exception occurred: " << e << std::endl;
+        }
     }
     else {
         try {
@@ -618,7 +224,7 @@ int main(int argc, char** argv)
             processor.run();
         }
         catch (int e) {
-            std::cout << "An exception occurred. Exception Nr. " << e << '\n';
+            std::cout << "An exception occurred. Exception Nr. " << e << std::endl;;
         }
     
         return 0;
